Refatorado imports dos teste e configurado paths con ts-jest em jest.config.ts

+++ src/modules/accounts/useCases/updateUserAvatar/UpdateUserAvatarUseCase.ts
@@ -1,7 +1,8 @@
 import { inject, injectable } from 'tsyringe';
 
-import { deleteFile } from '../../../utils/file';
-import { IUsersRepository } from '../repositories/IUsersRepository';
+import { AppError } from '@errors/AppError';
+import { IUsersRepository } from '@modules/accounts/repositories/IUsersRepository';
+import { deleteFile } from '@utils/file';
 
 interface IRequest {
   user_id: string;
@@ -17,6 +18,10 @@ class UpdateUserAvatarUseCase {
 
   async execute({ user_id, avatar_file }: IRequest): Promise<void> {
     const user = await this.usersRepository.findById(user_id);
+
+    if (!user) {
+      throw new AppError('User not exists', 400);
+    }
 
     if (user.avatar) {
       await deleteFile(`./tmp/avatar/${user.avatar}`);
